Added testcase for an XSS Attack, and slight restructuring of tests.
