+++ test/Chirp.RazorTest/End2EndTests/BadInteractionsTest.cs
@@ -1,75 +1,9 @@
-using Microsoft.Playwright;
+﻿using Microsoft.Playwright;
 
-namespace Chirp.RazorTest;
+namespace Chirp.RazorTest.End2EndTests;
 
-public class End2EndTest : PlaywrightPageTester
+public class BadInteractionsTest : PlaywrightPageTester
 {
-    private const string
-        DefaultEmail = "ropf@itu.dk",
-        DefaultUsername = "Helge",
-        DefaultPassword = "LetM31n!";
-    
-    [Test]
-    public async Task AnonymousUserNotLoggedIn()
-    {
-        await AssertNotLoggedIn();
-    }
-    
-    [Test]
-    public async Task CanRegisterAccount()
-    {
-        int id = Random.Shared.Next(1_000_000);
-        string email = $"testerson{id}@playwright.com";
-        string username = $"Playwright Testerson {id}";
-        
-        await Register(email, username, "Pa$$w0rd");
-        await AssertLoggedInAs(username);
-    }
-
-    [Test]
-    public async Task CanLogin()
-    {
-        await Login(DefaultUsername, DefaultPassword);
-        await AssertLoggedInAs(DefaultUsername);
-    }
-
-    [Test]
-    public async Task CanLogout()
-    {
-        await Login(DefaultUsername, DefaultPassword);
-        await AssertLoggedInAs(DefaultUsername);
-        
-        await ClickButton("logout");
-        await AssertNotLoggedIn();
-    }
-    
-    [Test]
-    public async Task CanCheep()
-    {
-        await Login(DefaultUsername, DefaultPassword);
-        
-        int id = Random.Shared.Next(1_000_000);
-        string cheep = $"Cheeping with playwright! {id}";
-        
-        await PostCheep(cheep);
-        await AssertContainsText("#messagelist", cheep);
-    }
-
-    [Test]
-    public async Task CheepsAppearOnOwnTimeline()
-    {
-        await Login(DefaultUsername, DefaultPassword);
-        
-        int id = Random.Shared.Next(1_000_000);
-        string cheep = $"Cheeping for my homies with playwright! {id}";
-        
-        await PostCheep(cheep);
-        
-        await ClickLink("my timeline");
-        await AssertContainsText("h2", $"{DefaultUsername}'s Timeline");
-        await AssertContainsText("#messagelist", cheep);
-    }
-
     [Test]
     public async Task LoginWithMissingFields()
     {
